Added Layers for Obstacles
When we place obstacles and fish, they may collide with each other and thi swill push the fish out of the tank. I added two layers to avoid this for now. It doesnt make sense because the fish will go though woods and stuff. No other solution comes to my mind for now.